feat: Pre-process the data
First, the data is split into a list of words and at the same time
punctuation signs are separated from actual words. Then two dictionaries
are created that translate back and forth between all the words in the
list and uniquely assigned integers. These dictionaries are then used
to transform the data into a list of integers.
The pre-processed data is saved in the "preprocess.p"-file.